simple test PKI and rcgen generator
Populate a tests/testsdata directory with a PEM encoded CA certificate,
a PEM encoded server certificate issued by the CA, and the PEM encoded
private key for the server certificate.

All of the above is generated by the `generate.rs` script using `rcgen`.
Run with cargo nightly (for cargo-script support):

  ./tests/testsdata/generate.rs